+++ Modules/FV0/CMakeLists.txt
@@ -7,9 +7,7 @@ target_sources(O2QcFV0 PRIVATE src/TH1ReductorLaser.cxx
                                src/DigitQcTask.cxx
                                src/CFDEffCheck.cxx
                                src/BasicPPTask.cxx
-                               src/OutOfBunchCollTask.cxx
-                               src/OutOfBunchCollCheck.cxx
-                               src/TriggerQcTask.cxx)
+                               src/OutOfBunchCollCheck.cxx)
 
 target_include_directories(
   O2QcFV0
@@ -33,9 +31,7 @@ add_root_dictionary(O2QcFV0
           include/FV0/DigitQcTaskLaser.h
           include/FV0/DigitQcTask.h
           include/FV0/BasicPPTask.h
-          include/FV0/OutOfBunchCollTask.h
           include/FV0/CFDEffCheck.h
-          include/FV0/TriggerQcTask.h
           include/FV0/OutOfBunchCollCheck.h
   LINKDEF include/FV0/LinkDef.h
   BASENAME O2QcFV0)
