+++ Modules/FV0/include/FV0/BasicPPTask.h
@@ -19,6 +19,7 @@
 
 #include "QualityControl/PostProcessingInterface.h"
 #include "QualityControl/DatabaseInterface.h"
+#include "CCDB/CcdbApi.h"
 
 #include "FV0Base/Constants.h"
 #include "DataFormatsFV0/ChannelData.h"
@@ -51,14 +52,18 @@ class BasicPPTask final : public quality_control::postprocessing::PostProcessing
   constexpr static std::size_t sNCHANNELS_PM = o2::fv0::Constants::nPms * o2::fv0::Constants::nChannelsPerPm;
 
  private:
+  std::string mPathGrpLhcIf;
   std::string mPathDigitQcTask;
   std::string mCycleDurationMoName;
+  std::string mCcdbUrl;
   int mNumOrbitsInTF;
 
   std::map<o2::fv0::ChannelData::EEventDataBit, std::string> mMapChTrgNames;
   std::map<int, std::string> mMapDigitTrgNames;
 
   o2::quality_control::repository::DatabaseInterface* mDatabase = nullptr;
+  o2::ccdb::CcdbApi mCcdbApi;
+
   std::unique_ptr<TGraph> mRateOrA;
   std::unique_ptr<TGraph> mRateOrAout;
   std::unique_ptr<TGraph> mRateOrAin;
@@ -74,8 +79,13 @@ class BasicPPTask final : public quality_control::postprocessing::PostProcessing
   std::unique_ptr<TCanvas> mRatesCanv;
   TProfile* mAmpl = nullptr;
   TProfile* mTime = nullptr;
+
+  // if storage size matters it can be replaced with TH1
+  // and TH2 can be created based on it on the fly, but only TH1 would be stored
+  std::unique_ptr<TH2F> mHistBcPattern;
+  std::unique_ptr<TH2F> mHistBcTrgOutOfBunchColl;
 };
 
 } // namespace o2::quality_control_modules::fv0
 
-#endif //QC_MODULE_FV0_BASICPPTASK_H
+#endif // QC_MODULE_FV0_BASICPPTASK_H
