--- Modules/FV0/include/FV0/OutOfBunchCollTask.h
@@ -1,60 +1,0 @@
-// Copyright 2019-2020 CERN and copyright holders of ALICE O2.
-// See https://alice-o2.web.cern.ch/copyright for details of the copyright holders.
-// All rights not expressly granted are reserved.
-//
-// This software is distributed under the terms of the GNU General Public
-// License v3 (GPL Version 3), copied verbatim in the file "COPYING".
-//
-// In applying this license CERN does not waive the privileges and immunities
-// granted to it by virtue of its status as an Intergovernmental Organization
-// or submit itself to any jurisdiction.
-
-///
-/// \file   OutOfBunchCollTask.h
-/// \author Sebastian Bysiak sbysiak@cern.ch
-///
-
-#ifndef QC_MODULE_FV0_OUTOFBUNCHCOLLTASK_H
-#define QC_MODULE_FV0_OUTOFBUNCHCOLLTASK_H
-
-#include "QualityControl/PostProcessingInterface.h"
-#include "QualityControl/DatabaseInterface.h"
-#include "CCDB/CcdbApi.h"
-
-#include "TList.h"
-#include "Rtypes.h"
-
-class TH1F;
-class TH2F;
-
-namespace o2::quality_control_modules::fv0
-{
-
-/// \brief PostProcessing task which finds collisions not compatible with BC pattern
-/// \author Sebastian Bysiak sbysiak@cern.ch
-class OutOfBunchCollTask final : public quality_control::postprocessing::PostProcessingInterface
-{
- public:
-  OutOfBunchCollTask() = default;
-  ~OutOfBunchCollTask() override = default;
-  void initialize(quality_control::postprocessing::Trigger, framework::ServiceRegistry&) override;
-  void update(quality_control::postprocessing::Trigger, framework::ServiceRegistry&) override;
-  void finalize(quality_control::postprocessing::Trigger, framework::ServiceRegistry&) override;
-  void configure(std::string, const boost::property_tree::ptree&) override;
-
- private:
-  std::string mPathDigitQcTask;
-  std::string mPathGrpLhcIf;
-  o2::quality_control::repository::DatabaseInterface* mDatabase = nullptr;
-  std::string mCcdbUrl;
-  o2::ccdb::CcdbApi mCcdbApi;
-  std::map<int, std::string> mMapDigitTrgNames;
-  // if storage size matters it can be replaced with TH1
-  // and TH2 can be created based on it on the fly, but only TH1 would be stored
-  std::unique_ptr<TH2F> mHistBcPattern;
-  std::unique_ptr<TH2F> mHistBcTrgOutOfBunchColl;
-};
-
-} // namespace o2::quality_control_modules::fv0
-
-#endif // QC_MODULE_FV0_OUTOFBUNCHCOLLTASK_H
